Migrated all table-unique database CRUD functions to individual table classes
