+++ src/utils/database/__init__.py
@@ -1,0 +1,3 @@
+"""
+    Empty file strictly here for importing Python modules from a subdirectory
+"""
