+++ src/utils/database/weekly_challenge_table.py
@@ -1,0 +1,82 @@
+"""
+Weekly Challenge table module
+"""
+
+from tinydb import TinyDB, where
+
+# Decorator for validating database insert methods
+def validate_insert(required_fields):
+    """
+    Python decorator function to validate objects before they are inserted
+    into the database.
+
+    Usage:
+        Before any method that inserts data into the database, you should
+        decorate the method as such:
+
+        @validate_insert(required_fields = TABLE_XXXX_FIELDS)
+
+        In the above, `TABLE_XXXX_FIELDS` should be the class constant variable
+        containing a list of the required fields for the table to be inserted
+        into.
+
+    Note:
+        This method does not prevent inserting duplicate values for unique
+        fields. The insert method body should still enforce uniqueness
+        where appropriate.
+    """
+
+    def decorator(func):
+        def wrapper(self, *args):
+            if (
+                isinstance(args[0], dict)
+                and all(field in args[0].keys() for field in required_fields)
+                and len(args[0].keys()) == len(required_fields)
+            ):
+                return func(self, *args)
+            return False  # Indicates validation failure
+
+        return wrapper
+
+    return decorator
+
+class WeeklyChallengeTable():
+    """
+    Table to hold all Weekly Challenges after being generated
+    """
+    TABLE_WEEKLY_CHALLENGE_FIELDS = ["id", "date"]
+    TABLE_WEEKLY_CHALLENGE = "Weekly_Challenge"
+
+    def __init__(self, database: TinyDB):
+        self.table = database.table(self.TABLE_WEEKLY_CHALLENGE)
+
+    @validate_insert(required_fields=TABLE_WEEKLY_CHALLENGE_FIELDS)
+    def insert(self, item: dict) -> bool:
+        """
+        "Inserts an item to the Weekly_Challenge table"
+        """
+
+        # Prevent duplicate ID's
+        if len(self.table.search(where("id") == item["id"])) == 0:
+            self.table.insert(item)
+            return True
+        return False
+
+    def load(self, challenge_id):
+        """
+        Loads an item by id from the Weekly_Challenge table
+        """
+        results = self.table.search(where("id") == challenge_id)
+        if len(results) == 0:
+            return None
+        return results[0]
+
+    def get_latest(self):
+        """
+        Loads the item with the most recent date property
+        """
+        results = self.table.all()
+        if len(results) == 0:
+            return None
+        sorted_results = sorted(results, key=lambda challenge: challenge["date"])
+        return sorted_results[-1]
